탭바 클릭 시 화면전환 방법 변경
