+++ GOOUT/Controller/InquiryByGradeViewController.swift
@@ -13,12 +13,6 @@ import DropDown
 class InquiryByGradeViewController : UIViewController, UITableViewDelegate, UITableViewDataSource {
     
     //MARK: - Properties
-    lazy var mainTabBarView = MainTabBarView()
-    
-    lazy var viewControllerView = UIView().then{
-        $0.backgroundColor = .white
-    }
-    
     lazy var dropLabelBtn = UIButton().then {
         $0.backgroundColor = .white
     }
@@ -115,6 +109,7 @@ class InquiryByGradeViewController : UIViewController, UITableViewDelegate, UITa
     lazy var earlyLeaveListHeader = EarlyLeaveListHeaderView().then {
         $0.backgroundColor = .rgb(red: 255, green: 243, blue: 243)
     }
+        
     // MARK: - Lifecycle
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -125,9 +120,6 @@ class InquiryByGradeViewController : UIViewController, UITableViewDelegate, UITa
     func configureUI(){
         view.backgroundColor = .white
         
-        layoutSetting()
-        
-        mainTabBarViewSetting()
         stateViewSetting()
         doNotGoViewSetting()
         tableViewSetting()
@@ -306,49 +298,6 @@ class InquiryByGradeViewController : UIViewController, UITableViewDelegate, UITa
         return 0
     }
     
-    // MARK: - myClassInquiryButtonClicked
-    @objc func myClassInquiryButtonClicked(sender:UIButton){
-        viewControllerView.isHidden = false
-        mainTabBarView.myClassInquiryButton.setImage(UIImage(named: "GOOUT_SelectedMyClassInquiryButtonImage"), for: .normal)
-        mainTabBarView.allClassInquiryButton.setImage(UIImage(named: "GOOUT_AllClassInquiryButtonImage"), for: .normal)
-    }
-    
-    // MARK: - allClassInquiryButtonClicked
-    @objc func allClassInquiryButtonClicked(sender:UIButton){
-        viewControllerView.isHidden = true
-        mainTabBarView.myClassInquiryButton.setImage(UIImage(named: "GOOUT_MyClassInquiryButtonImage"), for: .normal)
-        mainTabBarView.allClassInquiryButton.setImage(UIImage(named: "GOOUT_SelectedAllClassInquiryButtonImage"), for: .normal)
-    }
-    
-    //MARK: - addButtonClicked
-    @objc func addButtonClicked(sender:UIButton){
-        let nextVC = AddViewController()
-        present(nextVC, animated: true)
-    }
-    
-    // MARK: - layoutSetting
-    func layoutSetting(){
-        self.view.backgroundColor = .white
-        
-        self.view.addSubview(mainTabBarView)
-        self.view.addSubview(viewControllerView)
-        
-        viewControllerView.snp.makeConstraints { make in
-            make.top.equalToSuperview()
-            make.width.equalToSuperview()
-            make.bottom.equalToSuperview().offset(-self.view.frame.height/10.15)
-        }
-
-        // myClassInquiryViewController 화면전환 준비
-        let myClassInquiryViewController = InquiryForEachClassViewController()
-        self.addChild(myClassInquiryViewController)
-        myClassInquiryViewController.view.frame = viewControllerView.frame
-        
-        viewControllerView.addSubview(myClassInquiryViewController.view)
-        
-        viewControllerView.isHidden = true
-    }
-    
     func headerViewSetting(){
         outListHeader.addSubview(outListHeader.nameLabel)
         outListHeader.addSubview(outListHeader.gradeClassNumLabel)
@@ -388,60 +337,6 @@ class InquiryByGradeViewController : UIViewController, UITableViewDelegate, UITa
         doNotGoView.addSubview(doNotGoView.noStudentLabel)
         
         doNotGoView.doNotGoViewSetting(screenHeight: self.view.frame.height)
-    }
-    
-    // MARK: - mainTabBarViewSetting
-    func mainTabBarViewSetting(){
-        mainTabBarView.allClassInquiryButton.setImage(UIImage(named: "GOOUT_SelectedAllClassInquiryButtonImage"), for: .normal)
-        
-        mainTabBarView.myClassInquiryButton.addTarget(self, action: #selector(myClassInquiryButtonClicked(sender:)), for: .touchUpInside)
-        mainTabBarView.allClassInquiryButton.addTarget(self, action: #selector(allClassInquiryButtonClicked(sender:)), for: .touchUpInside)
-        mainTabBarView.addButton.addTarget(self, action: #selector(addButtonClicked(sender:)), for: .touchUpInside)
-        
-        mainTabBarView.snp.makeConstraints { make in
-            make.bottom.equalToSuperview()
-            make.width.equalToSuperview()
-            make.height.equalToSuperview().dividedBy(10.15)
-        }
-        
-        mainTabBarView.lineView.snp.makeConstraints { make in
-            make.top.equalToSuperview()
-            make.width.equalToSuperview()
-            make.height.equalTo(2)
-            make.centerX.equalToSuperview()
-        }
-        
-        mainTabBarView.addButton.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.width.equalToSuperview().dividedBy(14.37)
-            make.height.equalTo(mainTabBarView.addButton.snp.width)
-            make.top.equalTo(mainTabBarView.lineView.snp.bottom).offset(self.view.frame.height/58)
-        }
-        
-        mainTabBarView.allClassInquiryButton.snp.makeConstraints { make in
-            make.centerY.equalTo(mainTabBarView.addButton)
-            make.height.width.equalTo(mainTabBarView.addButton)
-            make.left.equalToSuperview().offset(self.view.frame.width/8.1)
-        }
-        
-        mainTabBarView.myClassInquiryButton.snp.makeConstraints { make in
-            make.centerY.equalTo(mainTabBarView.addButton)
-            make.height.width.equalTo(mainTabBarView.addButton)
-            make.right.equalToSuperview().offset(-self.view.frame.width/8.1)
-        }
-        
-        // MARK: - lineView gradient
-        let view = UIView(frame: CGRect(x: 0, y: 0, width: self.view.frame.width, height: 2))
-        let gradient = CAGradientLayer()
-
-        gradient.frame = view.bounds
-        gradient.colors = [UIColor(red: 104/255, green: 134/255, blue: 197/255, alpha: 1).cgColor, UIColor(red: 255/255, green: 173/255, blue: 172/255, alpha: 1).cgColor]
-        gradient.locations = [0.0 , 1.0]
-        gradient.startPoint = CGPoint(x: 0.0, y: 1.0)
-        gradient.endPoint = CGPoint(x: 1.0, y: 1.0)
-
-        mainTabBarView.lineView.layer.insertSublayer(gradient, at: 0)
-        
     }
     
 }
